requirements.txt
python-dotenv
# Discord의 최신 음성 암호화 모드(aead_xchacha20_poly1305_*) 지원 버전으로 고정
py-cord[voice]==2.6.3
PyNaCl
yt-dlp
